dashboard/analyse_dashboard/analyse/requirements.unexpanded.txt
pandas==0.24.2
firestore==0.0.8
google-cloud-secret-manager==1.0.0
numpy==1.16.4
xlrd==1.0.0
gcsfs==0.6.2
gobits==0.0.6
plotly==4.9.0
pyyaml==5.3.1
featuretoggles==1.0.1
matplotlib==3.2.1
mysql-connector-python==8.0.22
PyMySQL==0.10.1
SQLAlchemy==1.3.20
regex==2020.11.13